+++ src/jsTest/kotlin/com/ditchoom/socket/ReadStats.kt
@@ -3,7 +3,7 @@ package com.ditchoom.socket
 import kotlinx.coroutines.asDeferred
 import kotlin.js.Promise
 
-actual suspend fun readStats(port: UShort, contains: String): List<String> {
+actual suspend fun readStats(port: Int, contains: String): List<String> {
     if (TcpPortUsed.check(port.toInt(), "127.0.0.1").asDeferred().await()) {
         return listOf("TCP CHECK FAIL PORT: $port")
     }
